+++ vibrant-maven-plugin/src/main/java/io/teknek/deliverance/vibrant/VibrantMojo.java
@@ -1,14 +1,12 @@
 package io.teknek.deliverance.vibrant;
 
 import com.codahale.metrics.MetricRegistry;
-import com.google.j2objc.annotations.Property;
 import io.teknek.deliverance.DType;
 import io.teknek.deliverance.generator.GeneratorParameters;
 import io.teknek.deliverance.generator.Response;
 import io.teknek.deliverance.model.AbstractModel;
 import io.teknek.deliverance.model.ModelSupport;
 import io.teknek.deliverance.safetensors.fetch.ModelFetcher;
-import io.teknek.deliverance.safetensors.prompt.PromptContext;
 import io.teknek.deliverance.safetensors.prompt.PromptSupport;
 import io.teknek.deliverance.tensor.KvBufferCacheSettings;
 import io.teknek.deliverance.tensor.TensorCache;
@@ -27,7 +25,7 @@ import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
+import java.nio.file.StandardOpenOption;
 import java.util.List;
 import java.util.UUID;
 
@@ -39,9 +37,6 @@ public class VibrantMojo extends AbstractMojo {
 
     @Parameter(property = "outputDirectory", defaultValue = "${project.build.directory}/generated-sources/vibrant")
     private File outputDirectory;
-
-    @Parameter(property = "overwrite", defaultValue = "false")
-    private boolean overwrite;
 
     @Parameter
     private List<VibeSpec> vibeSpecs;
@@ -59,14 +54,25 @@ public class VibrantMojo extends AbstractMojo {
             spec.setGenerateTo("generated-source");
         }
         if ("generated-source".equalsIgnoreCase(spec.getGenerateTo())) {
-            boolean made = outputDirectory.mkdirs();
+            boolean ignore = outputDirectory.mkdirs();
             specDir = new File(outputDirectory, spec.id);
-            boolean made2 = specDir.mkdirs();
+            boolean ignore2 = specDir.mkdirs();
             project.addCompileSourceRoot(specDir.getAbsolutePath());
-            System.out.println("spec target" + specDir);
+            getLog().info("Generated source directory: " + specDir.getAbsolutePath());
+        } else if ("generated-test-source".equalsIgnoreCase(spec.getGenerateTo())) {
+            boolean ignore = outputDirectory.mkdirs();
+            specDir = new File(outputDirectory, spec.id);
+            boolean ignore2 = specDir.mkdirs();
+            project.addTestCompileSourceRoot(specDir.getAbsolutePath());
+            getLog().info("Generated test-source directory: " + specDir.getAbsolutePath());
         } else if ("existing-source".equalsIgnoreCase(spec.getGenerateTo())) {
             List<String> compileSourceRoots = project.getCompileSourceRoots();
             specDir = new File(compileSourceRoots.get(0));
+            getLog().info("Using source directory: " + specDir.getAbsolutePath());
+        } else if ("existing-test-source".equalsIgnoreCase(spec.getGenerateTo())) {
+            List<String> compileSourceRoots = project.getTestCompileSourceRoots();
+            specDir = new File(compileSourceRoots.get(0));
+            getLog().info("Using test-source directory: " + specDir.getAbsolutePath());
         } else {
             throw new MojoExecutionException("unsupported mode " + spec.getGenerateTo());
         }
@@ -82,20 +88,21 @@ public class VibrantMojo extends AbstractMojo {
                 .withTemperature(0.2f)
                 .withSeed(99998), (s1, f1) -> {
         });
-        System.out.println("k: " + k.responseText);
+        getLog().debug("inference engine reply: " + k.responseText);
         Path p = Path.of(specDir.toURI());
         Path child = p.resolve("raw.txt");
         JavaResponseTransformer t = new JavaResponseTransformer();
-        t.transform(k.responseText, p);
+        t.transform(k.responseText, p, spec);
         try {
-            Files.writeString(child, k.responseText);
+            if (spec.isOverwrite()){
+                Files.writeString(child, k.responseText);
+            } else {
+                Files.writeString(child, k.responseText, StandardOpenOption.CREATE_NEW);
+            }
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }
-
     }
-
-
 
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
@@ -104,7 +111,7 @@ public class VibrantMojo extends AbstractMojo {
         }
         boolean anyEnabled = vibeSpecs.stream().anyMatch(VibeSpec::isEnabled);
         if (!anyEnabled){
-            System.out.println("no vibespecs are enabled");
+            getLog().info("No VibeSpec are found or enabled");
             return;
         }
 
@@ -114,7 +121,7 @@ public class VibrantMojo extends AbstractMojo {
         TensorCache tensorCache = new TensorCache(mr);
         DType working = DType.valueOf(modelConfig.getWorkingMemType());
         DType quantized = DType.valueOf(modelConfig.getQuantizedMemType());
-        debudConfig();
+        debugConfig();
         NativeSimdTensorOperations operation = new NativeSimdTensorOperations(new ConfigurableTensorProvider(tensorCache).get());
         try (AbstractModel m = ModelSupport.loadModel(f, working, quantized, new ConfigurableTensorProvider(operation),
                 new MetricRegistry(), tensorCache, new KvBufferCacheSettings(true), fetch)) {
@@ -124,11 +131,10 @@ public class VibrantMojo extends AbstractMojo {
         }
     }
 
-    private void debudConfig(){
+    private void debugConfig(){
         System.out.println("Project: " + project.getGroupId() + ":"
                 + project.getArtifactId() + ":" + project.getVersion());
         System.out.println("vibe specs: " + vibeSpecs);
-        System.out.println("overwrite: " + overwrite);
         System.out.println("modelConfig: " + modelConfig);
     }
 }
